fix, curso, cursos a userRegisteredComponents, se agregó leftMenu a Cursos
