+++ frontend/src/components/UserRegisteredComponents/Curso/Curso.jsx
