+++ frontend/src/components/UserRegisteredComponents/Cursos/Cursos.jsx
@@ -1,0 +1,39 @@
+import { React, useEffect } from 'react'
+import { useDispatch, useSelector } from "react-redux";
+// import { Link } from "react-router-dom"
+import Curso from "../Curso/Curso";
+import { getCursos } from '../../../redux/actions/actionCurso';
+import LeftMenu from '../LeftMenu/LeftMenu';
+
+export default function Cursos() {
+    const dispatch = useDispatch();
+    const cursos = useSelector((state)=>state.allCursos)
+
+    useEffect(()=>{
+        dispatch(getCursos())
+    }, [])
+
+    return (
+    <div>
+        <div>
+            <LeftMenu/>
+        </div>
+        <div>
+            <h1>Da tus primeros pasos.</h1>
+            <h3>Estos son nuestros cursos.</h3>
+            {cursos?.map((e,i)=>{
+                return(
+                <div key={i}>
+                
+                <Curso
+                    id={e?._id}
+                    name={e?.name}
+                    description={e?.description}
+                    />
+                </div>
+                )
+            })}
+        </div>
+    </div>
+    )
+}
